requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
PyYAML==6.0.2
python-multipart==0.0.20
